requirements.txt
streamlit==1.21.0
sqlmodel==0.0.6
sqlite3==3.38.5
edenai-python==1.3.0
streamlit-cookies-manager==0.1.0
